+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.5
+
+- Updated package.json metadata
+- Updated databases merge logic
+- Added Wikipedia source
+- Added experimental field
+- Fixed docs files
+- Fixed mdn source description field scraping
+
 ## 1.0.4
 
 - Updated README.md usage example